deprecate kvstoreclientinternal
Summary:
as title this is not being used except in test.
In test, we were using a subscribe model where we trigger callback once the update completes. In order to remove dependency on this. We simply use a synchronous method(it's a test anyway) to read from queue and directly call the callback once there is an update

Reviewed By: xiangxu1121

Differential Revision: D42623870

fbshipit-source-id: 2ac3cf0e38bc284220e5843351afa4b686625610