Implement kvStore inconsistency detection logic for multiple types of inconsistencies
Summary:
# background
KvStore, as the key component of Open/R's distributed database to exchange information in between, should guarantee the eventual consistency with full-sync and incremental flooding.

# inconsistency
However, with some corner cases, KvStore keys can have inconsistency due to various reasons. It will be very dangerous to leave different stores in inconsistent state, which violates the fundamental of decision making process across the network. We need a way to detect them and mitigate this.

# thoughts
KvStore previously did periodic sync to address this kind of issues by randomly choosing a node every min, which in majority of the times, are a waste of network bandwidth since the node doesn't know which node has inconsistencies. To make this detection more efficient, ttl update can be genuinely the way to check since it is advertised and flooded across the network every quarter of the TTL expiration time. U can view TTL as the message containing all of the metadata needed for key consistency check.

# changes
This diff leverages the TTL update and implement the inconsistency detection mechanism by leveraging the information inside TTL update. High level speaking, it addresses:
1. originator sends ttl update but receiver does NOT have the key(ttl update ignored previously). Until next full-sync, two or mores nodes will be in inconsistency state;
2. originator sends ttl update but receiver has a different key version
3. originator sends ttl update but receiver has a different originatorId

Reviewed By: TangoRoxy

Differential Revision: D42950435

fbshipit-source-id: d8b0600568049ee640250173f35434f92b1d3f87