Add support for displaying the link metric increment
Summary:
As of now `breeze lm links` does not show that a link has a metric increment.
This makes it hard to figure out if a link is soft-drained.
`breeze openr neighbors` will show the increment, but that does not help if the link has no neighbors connected.

Also decided to sort the interfaces by slot/sub-slot/port rather than alphabetically.
The IPv6 link addresses are shown in the same line, it makes greping easier.

Reviewed By: xiangxu1121

Differential Revision: D45353888

fbshipit-source-id: 8c6d9c6debbf985de5450c758354bd4bf64c3dd7