do not remove neighborship if neighbor is just restarting
Summary:
# Overview
Original behavior of open/R:
To avoid partial adjacency, open/R has a logic to check if peer also consider itself as peer. If not, it would drop the neighborship with peer.

# Problem
This works great with graceful restart. However, consider the case of non-graceful restart.
Let's say we have simple topo A-B
B restart starts with a clean state. So it will send a message to A with empty adj. Now A will remove its neighborship with B until B do a full sync. Then they will be added again. This introduces unncessary flapping.

# Solution
We simply add a check to only do this neighborship removal if peer is not doing restart. How do we do that? Using seqNum. This seqNum is monotonically increasing. And if peer is restarting, this var would be smaller.

Reviewed By: TangoRoxy, xiangxu1121

Differential Revision: D45799983

fbshipit-source-id: e92f8592787af8f8cb215737067cfbc661e4ea1b