Update drained node selection logic
Summary:
Drained path selection is currently confusing
For example, if C is in the same area to A and B, which both (re)-originates P. Consider a case that:
```
A is drained, B is undrained but re-originates a prefix learnt from another drained device.
```
Then C will filter out the drained node A during best path selection before comparing the metric. So C would only choose B, even with a longer path.
{F1027272199}
Changed the logic so no softdrained node is filter, instead it updates the metrics so A and B would have the same drain metrics.

Reviewed By: xiangxu1121

Differential Revision: D46567908

fbshipit-source-id: fe6d96db07fcf53754819d3ee8acd175d56d0f99