Remove legacy thrift-py usage from utils
Summary:
As titled.

It is very complicated instance check between `thrift-py` and `thrift-python` :(  Given the fact that we have finished the majority of thrift-python migration, we should remove the legacy code to make our life easier. It is hard to simplify/consolidate logic if we have heavy burdens to carry over.

Reviewed By: shih-hao-tseng

Differential Revision: D47900826

fbshipit-source-id: 401dfa9e3b438b15f08a30cf5bc3d3fa71dc4511