onboard to serviceframework
Summary:
onboard open/R to serviceframework. Note that we only onboard the default thrift service, but the non-default one is left un-touched.

In order to use serviceframework, which is not open sourced, we create a new Main.cpp file under facebook directory. For de-duplication, we also create a MainUtil.cpp file so that we put common Main method there. Right now there are two: `startEventBase`, `waitForFibService`.

Reviewed By: xiangxu1121

Differential Revision: D47933046

fbshipit-source-id: 6068815696ec81d01b889a75016c98a1920640be