Remove fb303 value of decision.skipped_unicast_route
Summary: When we introduced the feature knob of IPv4 route with IPv6 next hops, we also introduced `decision.skipped_unicast_route` counter, which is a fb303 value. However, once we have rolled out the `v4_over_v6` feature and it has been deployed everywhere, we do not need to monitor this `decision.skipped_unicast_route` anymore.

Reviewed By: xiangxu1121

Differential Revision: D53075672

fbshipit-source-id: 47c335b5b88d5b8d137306b948bf6e0a097bf30b