Add hack namespace to thrift definitions
Summary: In order to make NHS calls from WWW, we need to sync its thrift dependencies. Some of those dependencies include the openr thrift files included in this diff. There are name collisions between the thrift definitions and existing Hack types in www so we need to add a hack namespace to the thrift definition.

Reviewed By: ashutoshgrewal

Differential Revision: D53244247

fbshipit-source-id: 44cd0f7c897b292cf4305a6a825aff6e959cca6f