KvStore flaky test due to underministic sleep
Summary:
Using sleep is a cause of flakyness of a lot of tests and is not recommended.
Linter would warn not use it.

however, we've found a cool way to defeat that by using scheduleTimeout() :)

When sleep or scheduleTimeout is used, the test may fail or pass depending on how long it takes to execute that test.

**Solution**

Wait for an event to ensure that something expected is received instead of waiting for random time and hoping that event has happened.

Reviewed By: xiangxu1121

Differential Revision: D53691681

fbshipit-source-id: e6972f5351e4b9ae30e9fb3a546043d9a03dbf2d