Add a record to keep link change (their status and timestamp) in AdjacentDB.
Summary:
We are trying to understand how long it takes for an up/down link event to be propagated along the network.

It's important to measure this because EBB LspAgent reacts to link changes. For example, if LspAgent sees a down link, it will switch to a new path to avoid traffic loss.

A timestamp of an up/down link event is kept in LinkEvents field in AdjacencyDB (similarly as we do for "PerfEvents") when Netlink in OpenR's LinkMonitor reports link event. You can ask why we don't use "PerfEvents"? It's because PerfEvents' data structure consists of string, which is not suitable for storing link event information.

LinkEvents list is quite small because in EBB we use aggregate ports. Typically a router has less than 20 port channels.

Reviewed By: xiangxu1121

Differential Revision: D53256463

fbshipit-source-id: 9d516985043dcd07093376a86f6a9a2c353a04cc