Follow-up T180252295: Allow same initial and max backoff for kvstore peer sync
Summary:
In a follow-up to T180252295. We want to allow kvstore peer sync backoffs configurable to same initial and max backoff value. Existing code had an assert at places and does not allow. Changes in this diffs,
- Remove assert checking equality of 2 values
- Relax configuration parameter check to allow max backoff to be same value as initial backoff
- Add Exponential backoff API unit tests
- Add kvstore peer sync unit tests

Not covered in this diffs: moving to friend class for use of processThriftSuccess/processThriftFailure APIs from test code.

Reviewed By: xiangxu1121

Differential Revision: D55074731

fbshipit-source-id: 2e7312c2d12f0988115d0a3fd9ec18860688946a