Adjust the EBB open/r config regex tests to accommodate EbLocal change
Summary: Because the MPLS labels 0-15 are reserved, naming for Eb-Local port-channels has been changed from Port-Channel{1 digit} to Port-Channel1000{1 digit} so that we can still use the ID of the Port-Channel as the MPLS label for the static MPLS routes. E.g., instead of Port-Channel1, the name will be Port-Channel10001. This diff adjusts the test for open/r regex accordingly.

Reviewed By: xiangxu1121, ngocdmfb

Differential Revision: D56543807

fbshipit-source-id: af939d9a916a67df3ba9be324139b0741d3577ef