change python3 to fbpython
Summary:
there is a bunch of scripts that uses python3 to run, this means that they are at the mercy of whoever controll their path, in linux, this can be system python3, but can also be platform python (and probably it is both), and in windows in particular it abuses c:\tools\fb-python\python3.exe. fbpython is the universal way to run python at the company that chooses the right platform python, but also provides monitoring and observability.

this scripts generates jobs like

https://www.internalfb.com/sandcastle/job/18014399781484898/

that will stop working once we remove c:\tools\fb-python\python3*.exe

Reviewed By: fried

Differential Revision: D56896564

fbshipit-source-id: d911fdaf6750635adb05b096f0522603baf47bcc