Deshim TestUtil in openr
Summary:
The following rules were deshimmed:
```
//folly/experimental:test_util -> //folly/testing:test_util
```

The following headers were deshimmed:
```
folly/experimental/TestUtil.h -> folly/testing/TestUtil.h
```

This is a codemod. It was automatically generated and will be landed once it is approved and tests are passing in sandcastle.
You have been added as a reviewer by Sentinel or Butterfly.

p:detuv.openr

Reviewed By: praihan

Differential Revision: D58833743

fbshipit-source-id: f5aaee1d3dfb13e7551c355becf8a78a2884132b