IOBuf moveToFbString().toStdString() -> toString()
Summary:
Constructing an `fbstring` from an `IOBuf` and then converting that to `std::string` is inefficient: if the `IOBuf` is chained, that can end up copying the data twice. `toString()` does exactly one copy.

This codemod replaces `moveToFbString().toStdString()` chained calls into `toString()`. If a `clone*()` call is part of the chain, it is removed, since `toString()` is non-destructive.

Note that this codemod is not exactly semantic-preserving because `moveToFbString()` is destructive and `toString()` is not, but nothing reuses `IOBuf` since it would be a pointless operation, so this should be safe.

Differential Revision: D58700808

fbshipit-source-id: ef07e6238005d088b022f077a473337b02907118