Expose ODS counters indicating number of KvStore peers in each PeerState
Summary:
# Changes
This diff includes changes to the OpenR KvStore module to publish counters for the number of KV store peers in each state (`IDLE`, `SYNCING`, and `INITIALIZED`). It also includes changes to the KvStoreUtil module to add a function to get the area type from a given area name and to publish counters for the number of peers in each state. Unit tests are introduced to cover both introduced functions below.

Reviewed By: xiangxu1121

Differential Revision: D60267079

fbshipit-source-id: d5c52d7e145320aba03de0a29ce73f008eb3e9a5