move openr spark fuzz test into team folder
Summary:
We are deprecating the /lionhead/harnesses/ directory. Please help us by relocating these harnesses out of this folder and into a new directory owned by the the owner of the code under test instead. This diff is just a proposed move based on our best guess for a good location.
Why:
While aggregating all harnesses in fbcode/security/lionhead/harnesses made sense in early days, the boom in number of harnesses has made it such that getting useful signal out of our contbuild has become unmaintainable for our tooling team. Since teams owning the software under test are better equipped to understand changes to their harnesses, it stands to reason that they will also be better equipped to house their own harnesses as opposed to potentially letting the harnesses fail silently on our end. We will still be reachable for continued support through the Lionhead Q&A group.
Action:
Just stamp if the location looks good : )
If you know a better location please commandeer this diff and move the harness to the correct location or let us know a better place.

Reviewed By: fbeqv

Differential Revision: D63138455

fbshipit-source-id: b63d711f3fb6eacfb37a702a1fe61bc1572a0b8e