Remove unused-variable in openr/link-monitor/tests/LinkMonitorTest.cpp
Summary:
LLVM-15 has a warning `-Wunused-variable` which we treat as an error because it's so often diagnostic of a code issue. Unused variables can compromise readability or, worse, performance.

This diff either (a) removes an unused variable and, possibly, it's associated code or (b) qualifies the variable with `[[maybe_unused]]`.

 - If you approve of this diff, please use the "Accept & Ship" button :-)

Reviewed By: xiangxu1121

Differential Revision: D64279121

fbshipit-source-id: ce2e3e61eb45e70b7e05b842a271cb2bbb1837bf