(red-diff) Cleanup enableSegmentRouting flag and corresponding logic
Summary:
As titled. There are a bunch of stale code inside Open/R with the segment routing support use cases, which has no longer active in any of the production flavor. Deprecate this for code simplicity.

More cleanup is coming to remove the footprint of segment routing use cases inside SpfSolver

Reviewed By: shih-hao-tseng

Differential Revision: D64424576

fbshipit-source-id: 3c883b88848f44d95f09acfdd69fbf1f846763b4