Fix the localPrefixConsidered check
Summary:
The check `myNodeName == prefixNode` is not enough to identify a local prefix. A prefix is local if
- myNodeName == prefixNode
- the area_stack is empty (not received from other areas)

This diff fixes the check.

This would lead to wrong calculation when
1. a locally originated prefix by switch A is also originated by others (say switch B)
2. the same prefix received from others (e.g., B) would be injected into a different area by switch A.

For instance, in DCType-F, we have some policy such as "typef_ild_fsw_mpod_level_infra" where
https://www.internalfb.com/code/configerator/[2cea48cc6701b6ed15170c4e055487d0bfb8936b]/raw_configs/neteng/ibr/intent/rpl/default_ild_typef_openr.rpl?lines=181-186

Normally, SSW (say, A) would inject to EXO_SPLICE, which would not inject to the neighboring MEGA_POD. Remote SSW (e.g., B) would not be preferred.

If the local SSW (A) is drained, the remote SSW (B) would be preferred and A would inject the prefix to MEGA_POD of A. In that case, A does not originate the prefix but it still has the prefix injected into MEGA_POD as a result of redistribution. Checking if "myNodeName == prefixNode" only does not give the correct signal here.

Reviewed By: xiangxu1121

Differential Revision: D64280631

fbshipit-source-id: c30c4b9050513c4598fe9886c37093d45a1bb3ed