update openr opensource build on github to bigger capacity machines
Summary:
https://docs.github.com/en/actions/using-github-hosted-runners/using-larger-runners/about-larger-runners#specifications-for-general-larger-runners

Currently builds on github are running only on 14GB machines which is not enough for the build. The builds have been failing for long due to disk space.

suggestion from opensource support:
https://fb.workplace.com/groups/osssupport/permalink/27367197886235467/

Specifying -8core towards the end hints github runners to pick up higher storage as shown in the link above.

Reviewed By: shih-hao-tseng

Differential Revision: D65950069

fbshipit-source-id: c0b9e13692bf644ce2110917cd8848a9df1dfa02