Validate if a different label can find runners
Summary:
Label specified in fix D65950069 does not work for github to find runner.

suggestion from open-source team is to try a different label. This is a quick change to first test and confirm. If it works, we will come back to re-work some code to generate the required label

Reviewed By: shih-hao-tseng

Differential Revision: D66217823

fbshipit-source-id: 54b5981b0c7670dd76d7603322877fd28aeedee1