correct netlink socket recv read buffer size
Summary:
netlink socket is set to send/rcv buffer size of 1MB. However, when socket is read with recv(), the buffer size it is reading is only 4KB which is much smaller than 1MB.

This causes reading of truncated data received via recv() call. Luckily, for EOS it works out because only interfaces (port-channel) we are interested in are sent by kernel in first 4KB size of the buffer.

On JunOS, it fails because port-channel interfaces are not sent out in first 4KB buffer size. When this recv() buffer size is increased to 64KB or 32KB, port-channel interfaces are read correctly in netlink notifications.

Reviewed By: xiangxu1121

Differential Revision: D66438614

fbshipit-source-id: 57c138df557db925c270dd17caf3356d09292d23