remove EQUAL_NE check
Summary:
The exact line that is flaky is  EXPECT_NE(fullSyncDurationAfter, fullSyncDurationBefore); is flaky. I think the main reason why this is, is because the key we are getting "kvstore.thrift.full_sync_duration_ms.avg is an average. From my investigation what I've found is that fullSyncDurationBefore captures 2 values in that average, and fullSyncDurationAfter captures 1 value after that. These functions run so fast, that usually I see the delta being 0-2 delta. What that means is that there is a high chance, that these 2 values will be equal.

A common scenario I see: fullSyncDurationBefore will capture values 1 and 2, causing its average to be 1, and fullSyncDurationAfter will capture 1 more value of 0, causing the average to stay 1.

As a result of the key representing average and us capturing so little additions to this average that the average sometimes doesnt change, I think the best solution would just be to remove that line to fix this.

Discussed offline with xiangxu1121, that the best fix for this is to just remove the check.

Reviewed By: xiangxu1121

Differential Revision: D76017376

fbshipit-source-id: 65eac0b04673b38bbf5a24c282269b0fc9022a61