Use a mutable heap in ttlCountdownQueue
Summary:
We change the heap structure to `d_ary_heap`. This allows us to have a mutable heap that can utilize the `update()` function, so that we update existing entries in the heap rather than add additional ones.

We keep track of current entries in the heap via `ttlCountdownHandleMap_`, where the key is `(key, originatorId)`

There was also an issue with `kvstore.ttl_countdown_queue_size` where we were getting the sum of the size on each call of update/erase. Fixed it so we are setting the size now on each call.

~~As discussed with xiangxu1121 offline, I will test out the removal of ttlVersion and version in a subsequent diff.~~

Edit: Discussed with xiangxu1121... After further investigation, we have decided to keep ttlVersion and version because it does have a purpose. When we add ttl to infinite, we do not add it to the countdown queue. This means that when setting a key there could be a difference between the publication we have received, and the entry inside the ttlCountdownQueue_. ttlVersion and version are used here so that we know that entry inside ttlCountdownQueue_ is outdated.

Reviewed By: xiangxu1121

Differential Revision: D76831947

fbshipit-source-id: 30686e8ba07e6dbc3bfd2fca90d047536cb47304