fix flaky test MonitorTestFixture.ProcessCounterTest
Summary:
We update our counters in `updateProcessCounters()`, which runs in a `folly::AsyncTimeout`.

Because in the above function `process.cpu.pct` is set before `process.cpu.peak_pct`, it is possible that when we detect our counters in the test, `process.cpu.peak_pct` is not set yet. To fix this, we change our conditional to check the last counter that should be set in `updateProcessCounters()`.

Reviewed By: xiangxu1121

Differential Revision: D77335670

fbshipit-source-id: 38c6695e3ccdddb72fc62368da734042128234d9