remove rtt_us counter
Summary:
Remove fb303 counter that measures rtt_us. This is a per neighbor counter.
There is already an existing counter called rtt_latest, which seems more important to keep in terms of if we want to keep the historical trend of rtt since it always shows the latest rtt, while `rtt_us` only updates based on the stepdetector (large changes to prevent too much noise).
Current `rtt_us` can also be found via command `breeze spark neighbors`

This is part of an effort to remove uneeded counters because we are getting closer to the quota.

Reviewed By: xiangxu1121

Differential Revision: D77769836

fbshipit-source-id: 7dfdbf9b962855e93eb140286a2cc7f31edc11ba