Stream all loopbacks unconditionally
Summary:
This diff adds an option, `stream_all_loopbacks`, to `streamNextHopStatus` API. If the option is set, then `FibAgent` considers every loopback router as a registered nexthop and sends the status without requiring any registration from BGP. This essentially replicates open/r route table in bgp, but bgp can avoid the blocking registration calls when the a nexthop first appears.

When the stream is first opened, FibAgent schedules a route dump on the stream using a new FunctionScheduler. The FunctionScheduler can later be used for IAR code.

Differential Revision: D78229035

fbshipit-source-id: b0e83dfb3e1c207e8308065b3c35307201f9d7be