Sync GSC Thrift Changes from Configerator to Fbcode
Summary:
Synced changes made in Configerator to Fbcode using `configerator-thrift-updater` for the following thrift files:
* bgp_config_v2.thrift
* feature_tag_v2.thrift
* nsf_policy_v2.thrift

These files were previously duplicated and updated in Configerator as part of the effort to migrate GenericSwitchConfig (GSC) to thrift-python.

Note that OpenrConfigV2.thrift was not synced using configerator-thrift-updater, as it was manually created and copied from OpenrConfig.thrift in Fbcode, respectively. This is because Fbcode is its source of truth.

Additionally, modified Buck targets to ensure compatibility with the updated thrift files and to maintain correct dependencies.

Reviewed By: shih-hao-tseng

Differential Revision: D77874826

fbshipit-source-id: 34d0b0ee9bc0f388fc75bc2b99050c6e2a8c4538