Fix flaky test OpenrCtrlFixture.SubscribeAndGetKvStore by using ScopedServerInterfaceThread
Summary:
`SubscribeAndGetKvStore` is flaky because we are reading the publications from expectedDeltas possibly while it is being modified outside of the ClientBufferedStream. Basically, the objects being passed in by ref to the callback and modified by the callback will also face the same problem (specifically the atomic int `received`).

The callback in question:
https://www.internalfb.com/code/fbsource/[354988fa3738]/fbcode/openr/ctrl-server/tests/OpenrCtrlHandlerTest.cpp?lines=2630-2640

We are also discouraged from using this deprecated method to get the Client stream:
https://www.internalfb.com/code/fbsource/[354988fa3738]/fbcode/openr/ctrl-server/tests/OpenrCtrlHandlerTest.cpp?lines=2627-2628
https://www.internalfb.com/code/fbsource/[4f6fb13e5d8b]/fbcode/thrift/lib/cpp2/async/ServerStream.h?lines=84-91

To fix the flakiness of this test while also maintaining the original test intent, we will use `ScopedServerInterfaceThread` to get an `AsyncGenerator` that will give us the published updates. From there, we check each value from the generator in sequence with our expected update. We can see in the original test pastes that not all of the publications were getting verified; this should also be fixed (the test used to clean up and finish before all deltas were checked).

Note that we are still using the original `semifuture_subscribeAndGetAreaKvStores`, so we are still testing the same behavior, except without the callback logic.

For anyone who is interested in seeing the read-after-free failures and not all publications getting verified:
Batch 1: P1884854297
Batch 2: P1885757876
Batch 3: P1885764826

Differential Revision: D79201805

fbshipit-source-id: cf9d7f9f1d2f71aaf26e725d5f74133a63101896