8/n combine netlinksocket and netlink wrapper evb
Summary:
While adding netlink to main.cpp I realized that netlink socket and netlink wrapper were running on different EVBs, meaning they would run on different threads. This is not what we wanted.

This diff adds changes to support running both the wrapper and socket on the same evb.

More Information:
Rather than passing in netlinksocket as a parameter in the ctor of the wrapper, we create the socket during ctor init. This allows us use the Evb that we inherit from BgpModuleBase. This also means that we pass in a replicatequeue instead of just the rqueue to the wrapper, as while the wrapper itself only needs a rqueue, socket requires the replicatequeue.

As discussed in the comments, there was a potential deadlock from `folly::coro::blockingWait(folly::coro::co_awaitTry(`, so instead we call `co_await folly::coro::co_awaitTry(` instead. I added this logic in a later diff, but moving it to this diff instead.

Reviewed By: xiangxu1121

Differential Revision: D80752406

fbshipit-source-id: 00f1635bf81d44906078538731d5ac4abbbb8172