Thrift changes to get the FibState from the FibAgent
Summary:
As part of the Bgp++ Fib Integration changes we need an endpoint in the FibAgent that can show all the routes and NHGs programmed by the FibAgent.

Task: T239891086

Added thrift changes to get the FibsState which consists of all the routes and nexthop groups programmed by the FibAgent.

Differential Revision: D85987076

fbshipit-source-id: 6bba4b02df1d59c1776f768e80eefed6e0edc49c