introduce cpu avg counter
Summary:
Adding counter `process.cpu.pct.avg`.

This will give us an average CPU pct. This differs from `process.cpu.pct`, which only gives us a cpu pct of a snapshot of time (configured to be 5 seconds).

We will use this in test_kvstore_ttl, which seems to be prone to flakiness as a result of sudden spikes in cpu within a snapshot of time.

Reviewed By: shih-hao-tseng

Differential Revision: D88895675

fbshipit-source-id: cb4cc799ffbea9ffecdd78ae44bd1d8e3d8d5557